Remove duplicate upserts
